refactor: Photographer DTO와 Model 파일 분리
