+++ Trinap/Sources/Domain/Model/Photographer.swift
@@ -1,0 +1,21 @@
+//
+//  Photographer.swift
+//  Trinap
+//
+//  Created by Doyun Park on 2022/11/16.
+//  Copyright © 2022 Trinap. All rights reserved.
+//
+
+import Foundation
+
+enum TagType: String {
+    case all
+    case family
+}
+
+struct Photograhper: Codable {
+    let photograhperId, location, introduction: String
+    let tags, pictures: [String]
+    let pricePerHalfHour: Int
+    let possibleDate: [Date]
+}
